base: extend object_pool by lookup_and_lock
Add functionality to lookup an object and lock it. Additional the case is
handled that a object may be already in-destruction and the lookup will deny
returning the object.

The object_pool generalize the lookup and lock functionality in rpc_server
and makes it obsolete. The lookup_and_lock feature is used to avoid recently
occurred races and deadlocks in base*.